Hoist shared VSM/HEKI wire types and errors into litebox_common_lvbs
Introduce litebox_common_lvbs holding the VTL0<->VTL1 ABI: VSM function IDs,
VsmError/HypervCallError/VerificationError, HEKI page/patch/range structures,
module/kexec wire types, MemAttr, and the page/PRK/core constants. The platform
deletes its duplicate definitions (mshv/heki.rs wire types, mshv/error.rs,
the hvcall HypervCallError enum, the kexec/module structs in host/linux.rs,
local MAX_CORES/PRK_LEN) and imports them from litebox_common_lvbs directly.

Platform-coupled helpers stay platform-side (mem_attr_to_hv_page_prot_flags,
which produces the platform HvPageProtFlags). No logic changes; the VSM handlers,
dispatch, registry, and mem_integrity remain in the platform for a later PR.
