+++ Cargo.toml
@@ -4,6 +4,7 @@ members = [
     "litebox", 
     "litebox_common_linux",
     "litebox_common_optee",
+    "litebox_common_lvbs",
     "litebox_platform_linux_kernel",
     "litebox_platform_linux_userland",
     "litebox_platform_windows_userland",
@@ -28,6 +29,7 @@ default-members = [
     "litebox",
     "litebox_common_linux",
     "litebox_common_optee",
+    "litebox_common_lvbs",
     "litebox_platform_linux_kernel",
     "litebox_platform_linux_userland",
     "litebox_platform_windows_userland",
