+++ litebox_platform_lvbs/Cargo.toml
@@ -8,6 +8,7 @@ edition = "2024"
 bitflags = "2.9.0"
 litebox = { path = "../litebox/", version = "0.1.0" }
 litebox_common_linux = { path = "../litebox_common_linux/", version = "0.1.0" }
+litebox_common_lvbs = { path = "../litebox_common_lvbs/", version = "0.1.0" }
 litebox_util_log = { version = "0.1.0", path = "../litebox_util_log" }
 spin = { version = "0.10.0", default-features = false, features = [
   "spin_mutex",
