+++ litebox_platform_lvbs/src/host/linux.rs
@@ -3,8 +3,8 @@
 
 //! Linux Structs
 
-use crate::arch::MAX_CORES;
-use zerocopy::{FromBytes, Immutable, IntoBytes, KnownLayout};
+use litebox_common_lvbs::MAX_CORES;
+use zerocopy::{FromBytes, Immutable, KnownLayout};
 
 /// Context saved when entering the kernel
 ///
@@ -90,86 +90,4 @@ impl CpuMask {
             }
         }
     }
-}
-
-#[allow(non_camel_case_types)]
-pub type __be32 = u32;
-
-#[repr(u8)]
-pub enum PkeyIdType {
-    PkeyIdPgp = 0,
-    PkeyIdX509 = 1,
-    PkeyIdPkcs7 = 2,
-}
-
-/// `module_signature` from [Linux](https://elixir.bootlin.com/linux/v6.6.85/source/include/linux/module_signature.h#L33)
-#[repr(C)]
-#[derive(Debug, Clone, Copy, FromBytes, Immutable, KnownLayout)]
-pub struct ModuleSignature {
-    pub algo: u8,
-    pub hash: u8,
-    pub id_type: u8,
-    pub signer_len: u8,
-    pub key_id_len: u8,
-    _pad: [u8; 3],
-    sig_len: __be32,
-}
-
-impl ModuleSignature {
-    pub fn sig_len(&self) -> u32 {
-        u32::from_be(self.sig_len)
-    }
-
-    /// Currently, Linux kernel only supports PKCS#7 signatures for module signing and thus `id_type` is always `PkeyIdType::PkeyIdPkcs7`.
-    /// Other fields except for `sig_len` are set to zero.
-    pub fn is_valid(&self) -> bool {
-        self.sig_len() > 0
-            && self.algo == 0
-            && self.hash == 0
-            && self.id_type == PkeyIdType::PkeyIdPkcs7 as u8
-            && self.signer_len == 0
-            && self.key_id_len == 0
-    }
-}
-
-/// `kexec_segment` from [Linux](https://elixir.bootlin.com/linux/v6.6.85/source/include/linux/kexec.h#L82)
-#[repr(C)]
-#[derive(Debug, Clone, Copy, FromBytes, IntoBytes, Immutable, KnownLayout)]
-pub struct KexecSegment {
-    /// Pointer to buffer (stored as u64 since we don't dereference it)
-    pub buf: u64,
-    pub bufsz: u64,
-    pub mem: u64,
-    pub memsz: u64,
-}
-
-/// `kimage` from [Linux](https://elixir.bootlin.com/linux/v6.6.85/source/include/linux/kexec.h#L296)
-/// Note that this is a part of the original `kimage` structure. It only contains some fields that
-/// we need for our use case, such as `nr_segments` and `segment`, and
-/// are not affected by the kernel build configurations like `CONFIG_KEXEC_FILE` and `CONFIG_IMA_KEXEC`.
-#[repr(C)]
-#[derive(Debug, Clone, Copy, FromBytes, IntoBytes, Immutable, KnownLayout)]
-pub struct Kimage {
-    head: u64,
-    /// Pointer fields stored as u64 since we don't dereference them
-    entry: u64,
-    last_entry: u64,
-    start: u64,
-    control_code_page: u64, // struct page*
-    swap_page: u64,         // struct page*
-    vmcoreinfo_page: u64,   // struct page*
-    vmcoreinfo_data_copy: u64,
-    pub nr_segments: u64,
-    pub segment: [KexecSegment; KEXEC_SEGMENT_MAX],
-    // we do not need the rest of the fields for now
-}
-pub const KEXEC_SEGMENT_MAX: usize = 16;
-
-/// `list_head` from [Linux](https://elixir.bootlin.com/linux/v6.6.85/source/include/linux/types.h#L190)
-/// Pointer fields stored as u64 since we don't dereference them.
-#[derive(Clone, Copy, Debug, FromBytes, IntoBytes, Immutable, KnownLayout)]
-#[repr(C)]
-pub struct ListHead {
-    pub next: u64,
-    pub prev: u64,
 }
