+++ litebox_platform_lvbs/src/host/mod.rs
@@ -8,7 +8,7 @@ pub mod lvbs_impl;
 pub mod per_cpu_variables;
 
 pub use lvbs_impl::LvbsLinuxKernel;
-pub(crate) use lvbs_impl::{PRK_LEN, set_platform_root_key};
+pub(crate) use lvbs_impl::set_platform_root_key;
 
 #[cfg(test)]
 pub mod mock;
