+++ litebox_platform_lvbs/src/mshv/hvcall_vp.rs
@@ -14,13 +14,14 @@ use crate::{
         HV_PARTITION_ID_SELF, HV_VP_INDEX_SELF, HV_VTL_NORMAL, HV_VTL_SECURE, HVCALL_ENABLE_VP_VTL,
         HVCALL_GET_VP_REGISTERS, HVCALL_SET_VP_REGISTERS, HvEnableVpVtl, HvGetVpRegistersInput,
         HvGetVpRegistersOutput, HvInputVtl, HvSetVpRegistersInput, SegmentRegisterAttributeFlags,
-        hvcall::{HypervCallError, hv_do_hypercall, hv_do_rep_hypercall},
+        hvcall::{hv_do_hypercall, hv_do_rep_hypercall},
         vtl1_mem_layout::{
             PAGE_SIZE, VTL1_KERNEL_STACK_PAGE, VTL1_TSS_PAGE, get_address_of_special_page,
         },
     },
     serial_println,
 };
+use litebox_common_lvbs::HypervCallError;
 use x86_64::{
     PrivilegeLevel,
     structures::{gdt::SegmentSelector, tss::TaskStateSegment},
